+++ src/components/modals/PrimaryNameModal/PrimaryNameModal.tsx
@@ -101,7 +101,8 @@ function PrimaryNameModal({
   const { data: primaryNameData, isLoading: isLoadingPrimaryNameData } =
     usePrimaryName();
 
-  const [{ transactionData }, dispatchTransactionState] = useTransactionState();
+  const [{ transactionData, signing }, dispatchTransactionState] =
+    useTransactionState();
   const [, dispatchArNSState] = useArNSState();
 
   const [fundingSource, setFundingSource] = useState<FundFrom>('balance');
@@ -512,11 +513,12 @@ function PrimaryNameModal({
             </div>
           )
         }
-        onCancel={closeModal}
-        onClose={closeModal}
+        onCancel={!signing ? closeModal : undefined}
+        onClose={!signing ? closeModal : undefined}
         onNext={
           !isLoading &&
           !isLoadingCostDetail &&
+          !signing &&
           costDetail &&
           (workflow === PRIMARY_NAME_WORKFLOWS.REMOVE ||
             (!isInsufficientBalance && !isInsufficientSolForGas))